requirements/ci.txt
-r base.txt

coverage==4.5.1
flake8==3.5.0
factory-boy==2.11.1
pytest==4.1.1
pytest-django==3.4.5
pytest-cov==2.6.1
moto==1.3.7